Fixed Chunked Pedestal. Now should work as intended, giving sensible results compared to the previous version

+++ python/src/bind_ClusterFinder.hpp
@@ -42,13 +42,13 @@ void define_ClusterFinder(py::module &m, const std::string &typestr) {
 
         .def("push_pedestal_mean",
              [](ClusterFinder<ClusterType, uint16_t, pd_type> &self,
-                py::array_t<uint16_t> frame, uint32_t chunk_number) {
+                py::array_t<double> frame, uint32_t chunk_number) {
                  auto view = make_view_2d(frame);                 
                  self.push_pedestal_mean(view, chunk_number);
              })
         .def("push_pedestal_std",
              [](ClusterFinder<ClusterType, uint16_t, pd_type> &self,
-                py::array_t<uint16_t> frame, uint32_t chunk_number) {
+                py::array_t<double> frame, uint32_t chunk_number) {
                  auto view = make_view_2d(frame);
                  self.push_pedestal_std(view, chunk_number);
              })
